feature: fail PPAC if ApiToken was used already in the current day, move apitoken logic into separate package
