+++ services/ppac/src/main/java/app/coronawarn/datadonation/services/ppac/ios/verification/PpacProcessor.java
@@ -4,6 +4,7 @@ import app.coronawarn.datadonation.common.persistence.domain.ApiToken;
 import app.coronawarn.datadonation.common.persistence.domain.DeviceToken;
 import app.coronawarn.datadonation.common.protocols.internal.ppdd.PpacIos.PPACIOS;
 import app.coronawarn.datadonation.services.ppac.ios.client.domain.PerDeviceDataResponse;
+import app.coronawarn.datadonation.services.ppac.ios.verification.apitoken.ApiTokenService;
 import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,16 +31,15 @@ public class PpacProcessor {
   }
 
   /**
-   * Incoming data submission requests must be validated before further processing. This means that
-   * it must be ensured that the request was indeed coming from a valid CWA client. The first step
-   * is to validate the DeviceToken {@link DeviceToken} against the Apple DeviceCheck API. Valid
-   * DeviceToken's are then stored to prevent replay attacks. Second step is to validate the
-   * provided ApiToken {@link ApiToken} and to update the corresponding per-Device Data (if existing
-   * or creating a new one).
+   * Incoming data submission requests must be validated before further processing. This means that it must be ensured
+   * that the request was indeed coming from a valid CWA client. The first step is to validate the DeviceToken {@link
+   * DeviceToken} against the Apple DeviceCheck API. Valid DeviceToken's are then stored to prevent replay attacks.
+   * Second step is to validate the provided ApiToken {@link ApiToken} and to update the corresponding per-Device Data
+   * (if existing or creating a new one).
    *
    * @param authentication              TODO
-   * @param ignoreApiTokenAlreadyIssued flag to indicate whether the ApiToken should be validated
-   *                                    against the last updated time from the per-device Data.
+   * @param ignoreApiTokenAlreadyIssued flag to indicate whether the ApiToken should be validated against the last
+   *                                    updated time from the per-device Data.
    * @param scenario                    TODO
    */
   public void validate(PPACIOS authentication, final boolean ignoreApiTokenAlreadyIssued,
