+++ services/ppac/src/test/java/app/coronawarn/datadonation/services/ppac/ios/PpaIosIntegrationTest.java
@@ -53,7 +53,6 @@ public class PpaIosIntegrationTest {
 
   @BeforeEach
   void clearDatabase() {
-
     when(jwtProvider.generateJwt()).thenReturn("jwt");
   }
