+++ services/ppac/src/test/java/app/coronawarn/datadonation/services/ppac/ios/controller/IosControllerTest.java
@@ -25,7 +25,7 @@ import app.coronawarn.datadonation.services.ppac.config.PpacConfiguration;
 import app.coronawarn.datadonation.services.ppac.config.TestBeanConfig;
 import app.coronawarn.datadonation.services.ppac.ios.client.IosDeviceApiClient;
 import app.coronawarn.datadonation.services.ppac.ios.client.domain.PerDeviceDataResponse;
-import app.coronawarn.datadonation.services.ppac.ios.verification.ApiTokenAuthenticator;
+import app.coronawarn.datadonation.services.ppac.ios.verification.apitoken.ApiTokenAuthenticator;
 import app.coronawarn.datadonation.services.ppac.ios.verification.JwtProvider;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
